+++ service/utils_test.go
@@ -5,19 +5,13 @@
 package service // import "github.com/wabarc/wayback/service"
 
 import (
-	"context"
 	"net/url"
-	"os"
 	"reflect"
 	"strconv"
 	"testing"
-	"time"
 
 	"github.com/wabarc/helper"
-	"github.com/wabarc/logger"
-	"github.com/wabarc/wayback"
 	"github.com/wabarc/wayback/config"
-	"github.com/wabarc/wayback/reduxer"
 )
 
 func TestMatchURL(t *testing.T) {
@@ -112,39 +106,5 @@ func TestExcludeURL(t *testing.T) {
 				t.Fatalf(`Unexpected exclude URLs number, got %v instead of %v`, got, test.want)
 			}
 		})
-	}
-}
-
-func TestWayback(t *testing.T) {
-	defer helper.CheckTest(t)
-
-	// Don't wayback to any slot to speed up testing.
-	os.Clearenv()
-	os.Setenv("WAYBACK_ENABLE_IA", "false")
-	os.Setenv("WAYBACK_ENABLE_IS", "false")
-	os.Setenv("WAYBACK_ENABLE_IP", "false")
-	os.Setenv("WAYBACK_ENABLE_PH", "false")
-
-	parser := config.NewParser()
-	var err error
-	if config.Opts, err = parser.ParseEnvironmentVariables(); err != nil {
-		t.Fatalf("Parse environment variables or flags failed, error: %v", err)
-	}
-	logger.SetLogLevel(logger.LevelFatal)
-
-	u, _ := url.Parse("https://example.com/")
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-	defer helper.CheckContext(ctx, t)
-
-	urls := []*url.URL{u}
-	do := func(cols []wayback.Collect, rdx reduxer.Reduxer) error {
-		time.Sleep(3 * time.Second)
-		return nil
-	}
-	w := Wayback(ctx, urls, do)
-
-	if w == nil {
-		t.Fatal("Unexpected wayback exceeded")
 	}
 }
